+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # _Sharly Chess_ release notes
 
+## Tournaments
+
+- Distribute the players among the tournaments (3.6.0)
+
+## Players
+
+- Import players from CSV files (3.6.0)
+
 ## Pairings
 
 - Upgrade bbpPairings engine to v6.0.0 (3.6.0)